requirements.txt
httpx==0.27.0
supabase==2.4.3
openai>=1.30.0
python-dotenv==1.0.1
rich>=13.7.0
